refactor(map): route Map.razor interop through typed wrappers
37 of the 44 raw module.InvokeAsync sites in Map.razor now go through
the IMap*Js wrappers, dropping the per-site try/catch dance. The
remaining 7 lifecycle-bound sites (initMap, the 6 dispose-time
clear*, dispose) stay bare with explanatory comments -- the
wrappers' MarkDisposed gate would otherwise no-op the teardown.

The InvokeVoidSafeAsync helper is gone; both fire-and-forget
sites in HandleSettingsChanged go through the wrappers now. The
SetGuardZoneAsync wrapper signature took double for lookaheadMin
to match IAppSettings.GuardZoneLookaheadMinutes (was int, now matches).

+++ OnaPlotter/Services/Js/IMapControlsJs.cs
@@ -42,7 +42,7 @@ public interface IMapControlsJs
     /// Radius is in nautical miles, lookahead in minutes; warning
     /// factor is the ratio at which the inner amber ring sits relative
     /// to the outer red ring.</summary>
-    Task SetGuardZoneAsync(double radiusNm, int lookaheadMin, double warningFactor);
+    Task SetGuardZoneAsync(double radiusNm, double lookaheadMin, double warningFactor);
 
     /// <summary>One-shot pan to the given lat/lon at the current zoom
     /// level. Distinct from <see cref="SetFollowAsync"/>, which
